requirements.txt
openai
ollama
groq
google-generativeai
pydantic
importlib-metadata    
python-dotenv
argparse        
pyyaml